requirements.txt
networkx==2.3
tensorflow-gpu==2.0.0
numpy==1.16.4
jupyter==1.0.0
pandas==0.25.1
scipy==1.3.1
scikit-learn==0.21.3
pytest==5.2.0
absl-py==0.8.1